+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Autoscaler Change Log
 
 - Fix getDemandFrames usage - sum resources before packing them into equal intervals
 - Fix time passed to scalingOptimizer in predictPolicy (make sure no event is skipped)
+- Fix scaling result calculation (fix equation + consider pending workload)
 
 0.9.0 November 8, 2020
 ------------------------